float мед

+++ Assets/Scripts/Game/API/Models/Log.cs
@@ -5,5 +5,5 @@ public class Log
 	public string Comment { get; set; }
 	public string PlayerName { get; set; }
 	public string ShopName {  get; set; }
-	public Dictionary<string, int> ResourcesChanged { get; set; }
+	public Dictionary<string, float> ResourcesChanged { get; set; }
 }
